Learn-Service/pyproject.toml
[project]
name = "learn-service"
version = "0.1.0"
description = "Microservice for supporting learning: course posting, student/instructor features."
authors = ["Iradukunda Fils <Iradukundafils1@gmail.com>"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "django==5.1.7",
    "django-appconf==1.1.0",
    "django-countries==7.6.1",
    "django-environ==0.12.0",
    "django-extensions==3.2.3",
    "django-ffmpeg==0.1.7",
    "django-rosetta==0.10.1",
    "django-statici18n==2.6.0",
    "fonttools==4.58.0",
    "future==1.0.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "kiwisolver==1.4.8",
    "lxml==5.4.0",
    "matplotlib==3.10.1",
    "numpy==2.2.5",
    "packaging==25.0",
    "pillow==11.1.0",
    "pluggy==1.6.0",
    "polib==1.2.0",
    "psycopg2-binary==2.9.10",
    "pyparsing==3.2.3",
    "pypdf2==3.0.1",
    "pytest==8.3.5",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-gettext==5.0",
    "python-magic==0.4.27",
    "python-pptx==1.0.2",
    "requests==2.32.3",
    "six==1.17.0",
    "soxr==0.5.0.post1",
    "sqlparse==0.5.3",
    "typing-extensions==4.13.2",
    "urllib3==2.4.0",
    "xlsxwriter==3.2.3",
]
